Fetching Data v5
Successfully eliminated my global variable by utilizing the
AppDelegate. I’ve cleaned up the code and commented the best I can for
readability. This version is closer to production. The only thing left
to do is either implement a load swirl, or attempt to increase the
speed by moving it up on the multi-thread “importance” lise.