Fetching Data v6
Increased efficiency and also implemented a load view which disappears
when the data gets there. This might be as good as I make this. Of
course anyone can try to make this faster. Use this as a starting point
for grabbing data in other views. NOTE: All records when pulled form
the Cloud are stored in the object: grabbedRecord.swift. So that can be
used to populate a TableView full of records which is Task #2 for the
database team.