FailsSafe implemented
After extensive research and testing, I was able to recreate the
network failure that occurred during class. The gods were not on our
side because there was a relatively small window for the breakage that
occurred during the presentation. Regardless I’ve implemented a fail
safe and have cleaned up code and comments.