Fix for Airplane Mode
Because the incremented sends an Asynchronous call to the iCloud
server, the user can navigate away from the homepage and then get
locked into the goals or history tab. To combat this, I pop the user
back to the homepage, but in doing so, a bug occurred where viewDidLoad
tried to grab the records from today even though it was in airplane
Mode and it caused a crazy glitch. It seems to be working fine now.