Fix For Loss of Records
In an effort to fix the bug where when we grabbed today’s Records more
than once and our records effectively doubled, I asked Rob how to erase
a current array, he showed me how but I accidentally put it in with the
loop when the array was being appended. So on each record that was
returned it was being immediately erased. Moving the reset of the array
out of the loop stopped the madness. Another rough day of
troubleshooting in the books.