A few more fixes
The App Delegate is where we need to “grab” things from iCloud. for
instance, when the app launches, when the app comes back from the
background (user goes to Twitter and then comes back to our app).
Basically, all situations need to be accounted for so I’ve been
testing, using the app on my phone and attempting to account for all
the changes that need to occur.