Final Fix
Storyboard: Fixed constraints with dataLabel and MaxLabel so that
maximum will appear as larger text if a higher goal is selected. Remade
the constraints in the average section. Removed mostSmokedDay label as
it was unused.

GoalTabViewController: Fixed yearly maximum calculation.

HistoryTabViewController: Removed IBOutlet to mostSmokedDay label
(removed). Added IBOutlets to maxTitle, recordHigh, recordLow, and
timeSinceLastSmoke labels. These labels were filled with false data in
each history view. Added variables to hold the count for each day. HTML
build functions now return the count of the number of records placed in
the table. This now calculates the counts in viewWillAppear and updates
the totals correctly for each view. Changed getHourlyAverage function
to grab the current hour of the day and divide that by today’s total.
Several other various fixes.